change names and update tests
