+++ src/main/resources/application.conf
@@ -2,8 +2,8 @@ akka {
   loggers = ["akka.event.slf4j.Slf4jLogger"]
   logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
   loglevel = "DEBUG"
-  log-dead-letters = off
   stdout-loglevel = "OFF"
+  log-dead-letters = off
 }
 
 api {
